Use parser-owned this method parsing
