+++ docs/COMPLETION_AUDIT.md
@@ -4015,6 +4015,9 @@ and do not assume Phase 4 is ready without evidence.
 - Parser pattern keyword dispatch now parses `true`, `false`, and `_` through
   `ParserPatternKeyword` instead of raw identifier spelling guards. Covered by
   `parser_pattern_keywords_use_owned_keyword_enum` and parser pattern tests.
+- Parser `@this.defer(...)` dispatch now parses `defer` through
+  `ParserThisMethod` instead of a raw method spelling check. Covered by
+  `parser_this_methods_use_owned_method_enum` and defer tests.
 - Gated `.raise()` and `.await()` method recognition now routes through
   `GatedMethod` enum-owned spellings, covered by
   `typechecker_gated_methods_use_owned_action_enum`,
