+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -6,6 +6,7 @@ mod ci_configs;
 mod file_size;
 mod module_system;
 mod parser_enums;
+mod parser_keywords;
 mod removed_syntax;
 mod source_truth;
 mod typechecker_imports;
